+++ apps/ledger-live-mobile/src/families/index.ts
@@ -1,5 +1,4 @@
 export * from "./algorand";
-export * from "./aptos";
 export * from "./bitcoin";
 export * from "./cardano";
 export * from "./celo";