+++ libs/coin-modules/coin-sui/src/api/index.ts
@@ -6,6 +6,8 @@ import {
   getBalance,
   listOperations,
   lastBlock,
+  getBlock,
+  getBlockInfo,
   craftTransaction,
 } from "../logic";
 import type { SuiAsset } from "./types";
@@ -27,6 +29,8 @@ export function createApi(config: SuiConfig): AlpacaApi<SuiAsset> {
     estimateFees: estimate,
     getBalance,
     lastBlock,
+    getBlock,
+    getBlockInfo,
     listOperations: list,
   };
 }
@@ -46,38 +50,5 @@ async function list(
   address: string,
   pagination: Pagination,
 ): Promise<[Operation<SuiAsset>[], string]> {
-  return operationsFromHeight(address, pagination.minHeight);
-}
-
-type PaginationState = {
-  readonly heightLimit: number;
-  continueIterations: boolean;
-  apiNextCursor?: string;
-  accumulator: Operation<SuiAsset>[];
-};
-
-async function operationsFromHeight(
-  address: string,
-  minHeight: number,
-): Promise<[Operation<SuiAsset>[], string]> {
-  const state: PaginationState = {
-    heightLimit: minHeight,
-    continueIterations: true,
-    accumulator: [],
-  };
-
-  while (state.continueIterations) {
-    const option = state.apiNextCursor
-      ? { minHeight: minHeight, cursor: state.apiNextCursor }
-      : { minHeight: minHeight };
-    const [operations, nextCursor] = await listOperations(address, option);
-    const opsFromHeight = operations.filter(op => op.tx.block.height > minHeight);
-    state.accumulator.push(...opsFromHeight);
-    state.apiNextCursor = nextCursor;
-    // The API makes 2 calls "IN" and "OUT" so even if we started filtering a few elements, we may still have more txs to fetch on one of the two calls.
-    // We can only stop when we've filtered all txs.
-    state.continueIterations = nextCursor !== "" && opsFromHeight.length > 0;
-  }
-
-  return [state.accumulator, state.apiNextCursor ? state.apiNextCursor : ""];
+  return listOperations(address, pagination);
 }