+++ libs/ui/packages/react/src/pre-ldls/components/Tag/Tag.tsx
@@ -3,7 +3,7 @@ import { Text } from "../../../components";
 import { withTokens } from "../../libs";
 import styled from "styled-components";
 
-const Wrapper = styled.div`
+const Wrapper = styled.div<{ $spacing: "sm" | "md" }>`
   ${withTokens(
     "colors-surface-transparent-subdued-default",
     "colors-content-subdued-default-default",
@@ -11,7 +11,7 @@ const Wrapper = styled.div`
     "spacing-xxxs",
   )}
 
-  padding: 1px var(--spacing-xxxs);
+  padding: ${({ $spacing }) => ($spacing === "sm" ? "1px" : "2px")} var(--spacing-xxxs);
   border-radius: var(--radius-xs);
   display: inline-flex;
   background-color: var(--colors-surface-transparent-subdued-default);
@@ -20,13 +20,15 @@ const Wrapper = styled.div`
 
 export const Tag = ({
   textTransform = "none",
+  spacing = "sm",
   children,
 }: {
   textTransform?: string;
+  spacing?: "sm" | "md";
   children: ReactNode;
 }) => {
   return (
-    <Wrapper data-testid="tag">
+    <Wrapper data-testid="tag" $spacing={spacing}>
       <Text
         color="var(--colors-content-subdued-default-default)"
         fontSize="10px"