+++ libs/ledgerjs/packages/hw-app-celo/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @ledgerhq/hw-app-polkadot
 
+## 6.33.11
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/hw-app-eth@6.45.13
+
+## 6.33.11-next.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/hw-app-eth@6.45.13-next.1
+
+## 6.33.11-next.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/hw-app-eth@6.45.13-next.0
+
 ## 6.33.10
 
 ### Patch Changes